definitions: Add type Z24/Unit and make Z21/Void to be its instance
Before this change: Z21 was the Unit type (the type that has
exactly one instance, which is informationless but exists.
Like void in C++ etc.)

Some function calls returned Z21 in order to indicate that they
return no information. But Z21 was the Type!

What we needed was to return the instance of the type.

In order to not have to change too much code, we continue let
those functions return Z21, but we redefine Z21 to be the instance
of the Unit type, and the Unit type itself is moved to Z24.

(This change is not tested)

Bug: T287195
Change-Id: Id13dd8e58d46bf99c4e9b57ddeeec21b5105cd4e