Canonicalize and normalize should return Z5s instead of throwing errors
Utility methods 'canonicalize' and 'normalize' were changed to return a tuple of data and error. Only one of those elements must be a non-empty value. The distinction is necessary because its the functions can return Z5s or indicate that an error ocurred (also using a Z5). Tests were also changed to check if the correct error codes (Z5xx) were being returned.

Bug: T287886
Change-Id: I15f5c652421adb05ed147717b22595231c5d4c24