Fix overlap between Z542 and Z551
The following error types needed clarification:
* Z542: ZObject type mismatch: We are expecting a ZObject of a given
type and we receive one of a different type. (E.g. we receive a Z4 when
we are expecting a Z8)
* Z551: Schema type mismatch: Static validation error, we are expecting
a json object of a given type and receive that of another type. (E.g. we
find an array where we are expecting a string)

Fixed:
* Renaming of Z551 to avoid confusion
* Change again the key types for Z551 (need to be strings)

Bug: T303188
Change-Id: I8813106093fa9449bc10ae73a2be26e83b28f3ea