[BREAKING CHANGE] Drop legacy flags from canonicalize(), normalize(), and SchemaFactory
We no longer use ZObjects with raw arrays instead of Benjamin format
('withBenjamin', 'fromBenjamin', 'toBenjamin'), ZObjects with null
instead of a 'Z24' reference ('withVoid'), or legacy 'Z10' untyped
lists ('generically').

Doing this all at once, as otherwise we get errors for passing params
which are unused or reading values that are unset.

Change-Id: I34360c27408574ada6e410ece91930a4a54fe30f